change the price feed back to latest
fix dydx module test cases

+++ lib/keys.go
@@ -8,6 +8,6 @@ import (
 // The slices can be ordered lexicographically
 func Uint32ToKey(i uint32) []byte {
 	bytes := make([]byte, 4)
-	binary.BigEndian.PutUint32(bytes, uint32(i))
+	binary.BigEndian.PutUint32(bytes, i)
 	return bytes
 }
